Transforming the scan mean method into algorithm and creating the method to apply the algorithm on the raw data
